Continue Split files
