+++ MVVMSidekick/MVVMSidekick.Shared/PlatformPatches.cs
@@ -1,5 +1,42 @@
 ﻿
 
+#if SILVERLIGHT_5 || WINDOWS_PHONE_8 || WINDOWS_PHONE_7
+	/// <summary>
+        /// Class ConcurrentDictionary.
+        /// </summary>
+        /// <typeparam name="TK">The type of the t property.</typeparam>
+        /// <typeparam name="TV">The type of the tv.</typeparam>
+        public class ConcurrentDictionary<TK, TV> : Dictionary<TK, TV>
+        {
+            /// <summary>
+            /// Gets the or add.
+            /// </summary>
+            /// <param name="key">The key.</param>
+            /// <param name="factory">The factory.</param>
+            /// <returns>TV.</returns>
+            public TV GetOrAdd(TK key, Func<TK, TV> factory)
+            {
+                TV rval = default(TV);
+
+                if (!base.TryGetValue(key, out rval))
+                {
+                    lock (this)
+                    {
+                        if (!base.TryGetValue(key, out rval))
+                        {
+                            rval = factory(key);
+                            base.Add(key, rval);
+                        }
+
+
+                    }
+                }
+
+                return rval;
+            }
+        }
+#endif
+
 #if NET40
 
 using System.Collections;
