+++ MVVMSidekick/MVVMSidekick.Shared/ViewModels/BindableBase.cs
@@ -1,12 +1,586 @@
 ﻿using MVVMSidekick.EventRouting;
+using MVVMSidekick.Utilities;
 using System;
 using System.Collections.Generic;
 using System.ComponentModel;
+using System.Linq;
+using System.Linq.Expressions;
+using System.Reactive.Linq;
 using System.Runtime.Serialization;
 using System.Text;
 
 namespace MVVMSidekick.ViewModels
 {
+    /// <summary>
+    /// <para>Model type with detail subtype type parameter.</para>
+    /// <para>具有子类详细类型定义的model </para>
+    /// <example>
+    /// public class Class1:BindableBase&lt;Class1&gt;  {}
+    /// </example>
+    /// </summary>
+    /// <typeparam name="TSubClassType">Sub Type / 子类类型</typeparam>
+    [DataContract]
+    public abstract class BindableBase<TSubClassType> : BindableBase, INotifyDataErrorInfo where TSubClassType : BindableBase<TSubClassType>
+    {
+        protected static Dictionary<string, Func<TSubClassType, IValueContainer>>
+            _plainPropertyContainerGetters =
+              new Dictionary<string, Func<TSubClassType, IValueContainer>>(StringComparer.CurrentCultureIgnoreCase);
+
+        static BindableBase()
+        {
+
+        }
+
+        /// <summary>
+        /// The _plain property container getters
+        /// </summary>
+
+
+
+
+        /// <summary>
+        /// Initializes a new instance of the <see cref="BindableBase{TSubClassType}"/> class.
+        /// </summary>
+        public BindableBase()
+        {
+            Observable.FromEventPattern<DataErrorsChangedEventArgs>(
+                eh => this._ErrorsChanged += eh,
+                eh => this._ErrorsChanged -= eh)
+                .Subscribe(_ =>
+                {
+                    HasErrors = GetAllErrors().Any();
+                })
+                .DisposeWith(this);
+
+            //_BindableInstanceIdLocator(this).SetValueAndTryNotify( string.Format("{0}:{1}", this.GetType().Name, base._instanceIdOfThisType));
+        }
+
+
+        /// <summary>
+        /// Gets the bindable instance identifier.
+        /// </summary>
+        /// <value>The bindable instance identifier.</value>
+
+        public override string BindableInstanceId
+        {
+            get { return _instanceIdOfThisType.ToString(); }
+
+        }
+
+
+
+
+
+
+        /// <summary>
+        /// 清除值
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="property">The property.</param>
+        public void ResetPropertyValue<T>(Property<T> property)
+        {
+            if (property != null)
+            {
+                var oldContainer = property.Container;
+                if (oldContainer != null)
+                {
+
+
+                    property.Container = null;
+                    property.LocatorFunc(oldContainer.Model);
+                    oldContainer.SetValueAndTryNotify(property.Container.Value);
+                    property.Container = oldContainer;
+                }
+            }
+
+
+        }
+
+
+
+
+
+
+
+
+        /// <summary>
+        /// <para>Cast a model instance to current model subtype</para>
+        /// <para>将一个 model 引用特化为本子类型的引用</para>
+        /// </summary>
+        /// <param name="model">some bindable model/某种可绑定model</param>
+        /// <returns>Current sub type instance/本类型引用</returns>
+        public static TSubClassType CastToCurrentType(BindableBase model)
+        {
+            return (TSubClassType)model;
+
+        }
+        ///// <summary>
+        ///// <para>Type cache of container getter</para>
+        ///// <para>每个属性类型独占的一个专门的类型缓存。</para>
+        ///// </summary>
+        ///// <typeparam name="TProperty"></typeparam>
+        //protected static class TypeDic<TProperty>
+        //{
+        //	public static Dictionary<string, Func<TSubClassType, ValueContainer<TProperty>>> _propertyContainerGetters = new Dictionary<string, Func<TSubClassType, ValueContainer<TProperty>>>();
+
+        //}
+
+        /// <summary>
+        /// 根据索引获取属性值
+        /// </summary>
+        /// <param name="colName">Name of the col.</param>
+        /// <returns>属性值</returns>
+        public override object this[string colName]
+        {
+            get
+            {
+                var lc = GetOrCreatePlainLocator(colName, this);
+                return lc((TSubClassType)this).Value;
+            }
+            set
+            {
+
+                var lc = GetOrCreatePlainLocator(colName, this);
+                lc((TSubClassType)this).Value = value;
+            }
+        }
+
+        /// <summary>
+        /// Gets the or create plain locator.
+        /// </summary>
+        /// <param name="colName">Name of the col.</param>
+        /// <param name="viewModel">The view model.</param>
+        /// <returns>Func&lt;TSubClassType, IValueContainer&gt;.</returns>
+        private static Func<TSubClassType, IValueContainer> GetOrCreatePlainLocator(string colName, BindableBase viewModel)
+        {
+            Func<TSubClassType, IValueContainer> pf;
+            if (!_plainPropertyContainerGetters.TryGetValue(colName, out pf))
+            {
+                var p = new ValueContainer<object>(colName, viewModel);
+
+                Func<TSubClassType, ValueContainer<object>> tpf = _ => p;
+                pf = tpf;
+                _plainPropertyContainerGetters[colName] = pf;
+                //TypeDic<object>._propertyContainerGetters[colName] = tpf;
+            }
+            return pf;
+        }
+
+
+
+
+        //#if SILVERLIGHT_5 || WINDOWS_PHONE_8 || WINDOWS_PHONE_7
+        //			/// <summary>
+        //			/// The _plain property container getters
+        //			/// </summary>
+        //			protected static Dictionary<string, Func<TSubClassType, IValueContainer>>
+        //			 _plainPropertyContainerGetters =
+        //			 new Dictionary<string, Func<TSubClassType, IValueContainer>>(StringComparer.CurrentCultureIgnoreCase);
+        //#else
+        //			/// <summary>
+        //			/// The _plain property container getters
+        //			/// </summary>
+        //			protected static Dictionary<string, Func<TSubClassType, IValueContainer>>
+        //				_plainPropertyContainerGetters =
+        //				new Dictionary<string, Func<TSubClassType, IValueContainer>>(StringComparer.CurrentCultureIgnoreCase);
+        //#endif
+
+
+
+        /// <summary>
+        /// Gets the error.
+        /// </summary>
+        /// <value>The error.</value>
+        public override string ErrorMessage
+        {
+            get { return _ErrorMessageLocator(this).Value; }
+        }
+
+        /// <summary>
+        /// Sets the error.
+        /// </summary>
+        /// <param name="value">The value.</param>
+        protected override void SetErrorMessage(string value)
+        {
+            _ErrorMessageLocator(this).SetValue(value);
+        }
+        /// <summary>
+        /// 
+        /// </summary>
+        public void GenrateErrorMessage()
+        {
+            var sb = new StringBuilder();
+            OnGenrateErrorsMessage(GetAllErrors(), sb);
+            SetErrorMessageAndTryNotify(sb.ToString());
+        }
+
+        /// <summary>
+        /// Sets the error and try notify.
+        /// </summary>
+        /// <param name="value">The value.</param>
+        protected override void SetErrorMessageAndTryNotify(string value)
+        {
+            _ErrorMessageLocator(this).SetValueAndTryNotify(value);
+        }
+
+
+
+
+        #region Property string ErrorMessage Setup        
+        protected Property<string> _ErrorMessage = new Property<string> { LocatorFunc = _ErrorMessageLocator };
+        static Func<BindableBase, ValueContainer<string>> _ErrorMessageLocator = RegisterContainerLocator<string>(nameof(ErrorMessage), model => model.Initialize(nameof(ErrorMessage), ref model._ErrorMessage, ref _ErrorMessageLocator, _ErrorMessageDefaultValueFactory));
+        static Func<string> _ErrorMessageDefaultValueFactory = () => default(string);
+        #endregion
+
+
+
+
+
+
+        /// <summary>
+        /// 注册一个属性容器的定位器。
+        /// </summary>
+        /// <typeparam name="TProperty">The type of the property.</typeparam>
+        /// <param name="propertyName">属性名</param>
+        /// <param name="getOrCreateLocatorMethod">属性定位/创建方法 也就是定位器</param>
+        /// <returns>
+        /// 注册后的定位器
+        /// </returns>
+        protected static Func<BindableBase, ValueContainer<TProperty>> RegisterContainerLocator<TProperty>(string propertyName, Func<TSubClassType, ValueContainer<TProperty>> getOrCreateLocatorMethod)
+        {
+
+            //TypeDic<TProperty>._propertyContainerGetters[propertyName] = getOrCreateLocatorMethod;
+            _plainPropertyContainerGetters[propertyName] = getOrCreateLocatorMethod;
+            return o => getOrCreateLocatorMethod((TSubClassType)o);
+
+
+        }
+
+
+        /// <summary>
+        /// 根据属性名取得一个值容器
+        /// </summary>
+        /// <typeparam name="TProperty">The type of the property.</typeparam>
+        /// <param name="propertyName">属性名</param>
+        /// <returns>值容器</returns>
+        /// <exception cref="System.Exception">
+        /// Property Not Exists!
+        /// or
+        /// Property ' + propertyName + ' is found but it does not match the property type ' + type of(TProperty).Name + '!
+        /// </exception>
+        public ValueContainer<TProperty> GetValueContainer<TProperty>(string propertyName)
+        {
+            Func<TSubClassType, ValueContainer<TProperty>> containerGetterCreater;
+            Func<TSubClassType, IValueContainer> contPlanGetter;
+            if (!_plainPropertyContainerGetters.TryGetValue(propertyName, out contPlanGetter))
+            {
+                throw new Exception("Property Not Exists!");
+
+            }
+
+            containerGetterCreater = contPlanGetter as Func<TSubClassType, ValueContainer<TProperty>>;
+            if (containerGetterCreater == null)
+            {
+                throw new Exception("Property '" + propertyName + "' is found but it does not match the property type '" + typeof(TProperty).Name + "'!");
+            }
+
+            return containerGetterCreater((TSubClassType)(Object)this);
+
+        }
+
+        /// <summary>
+        /// 根据表达式树取得一个值容器
+        /// </summary>
+        /// <typeparam name="TProperty">The type of the property.</typeparam>
+        /// <param name="expression">表达式树</param>
+        /// <returns>
+        /// 值容器
+        /// </returns>
+        public ValueContainer<TProperty> GetValueContainer<TProperty>(Expression<Func<TSubClassType, TProperty>> expression)
+        {
+            var propName = MVVMSidekick.Utilities.ExpressionHelper.GetPropertyName<TSubClassType, TProperty>(expression);
+            return GetValueContainer<TProperty>(propName);
+
+        }
+
+
+
+
+        /// <summary>
+        /// 根据属性名取得一个值容器
+        /// </summary>
+        /// <param name="propertyName">属性名</param>
+        /// <returns>值容器</returns>
+        /// <exception cref="System.NotImplementedException"></exception>
+        public override IValueContainer GetValueContainer(string propertyName)
+        {
+            Func<TSubClassType, IValueContainer> contianerGetterCreater;
+            if (!_plainPropertyContainerGetters.TryGetValue(propertyName, out contianerGetterCreater))
+            {
+                this[propertyName] = null;
+                if (!_plainPropertyContainerGetters.TryGetValue(propertyName, out contianerGetterCreater))
+                {
+                    throw new NotImplementedException(string.Format("Current property \"{0}\" is not implemented", propertyName));
+                }
+            }
+            return contianerGetterCreater((TSubClassType)(Object)this);
+
+        }
+
+
+        /// <summary>
+        /// 根据属性名取得多个值容器
+        /// </summary>
+        /// <param name="propertyNames">The property names.</param>
+        /// <returns>值容器</returns>
+        public IValueContainer[] GetValueContainers(params string[] propertyNames)
+        {
+            return propertyNames.Select(pn => GetValueContainer(pn)).ToArray();
+
+        }
+
+        /// <summary>
+        /// 根据表达式树取得多个值容器
+        /// </summary>
+        /// <param name="expressions">The expressions.</param>
+        /// <returns>值容器</returns>
+        public IValueContainer[] GetValueContainers(params Expression<Func<TSubClassType, object>>[] expressions)
+        {
+
+            var names = expressions.Select(expression =>
+                  MVVMSidekick
+                    .Utilities
+                    .ExpressionHelper
+                    .GetPropertyName<TSubClassType>(expression)
+                ).ToArray();
+
+            return GetValueContainers(names);
+        }
+
+
+
+
+
+
+        /// <summary>
+        /// 获取所有属性名，包括静态声明和动态添加的
+        /// </summary>
+        /// <returns>System.String[].</returns>
+        public override string[] GetFieldNames()
+        {
+            return _plainPropertyContainerGetters.Keys.ToArray();
+        }
+
+
+        /// <summary>
+        /// 创建一个VM副本
+        /// </summary>
+        /// <returns>新引用</returns>
+        public TSubClassType Clone()
+        {
+            var x = (TSubClassType)Activator.CreateInstance(typeof(TSubClassType));
+            CopyTo(x);
+            return x;
+        }
+
+        /// <summary>
+        /// Copyrefs the specified source.
+        /// </summary>
+        /// <typeparam name="T"></typeparam>
+        /// <param name="source">The source.</param>
+        /// <param name="target">The target.</param>
+        static void Copyref<T>(T source, ref T target)
+        {
+
+
+            if (source == null)
+            {
+                target = source;
+                return;
+            }
+
+            var sourcetype = source.GetType().GetTypeOrTypeInfo();
+            if (sourcetype.IsValueType || source is string)
+            {
+                target = source;
+            }
+#if !(SILVERLIGHT_5 || WINDOWS_PHONE_8 || WINDOWS_PHONE_7 || NETFX_CORE)
+
+                else if (typeof(ICloneable).IsAssignableFrom(sourcetype))
+                {
+                    target = (T)((ICloneable)source).Clone();
+                }
+#endif
+            else if (typeof(System.Collections.IList).GetTypeOrTypeInfo().IsAssignableFrom(sourcetype))
+            {
+                var tarcol = target as System.Collections.IList;
+                var scol = source as System.Collections.IList;
+                if (tarcol == null)
+                {
+
+                    var newcol = sourcetype.IsArray ?
+                        Array.CreateInstance(sourcetype.GetElementType(), scol.Count) :
+                        System.Activator.CreateInstance(source.GetType(), new object[0]) as System.Collections.IList;
+
+
+                    tarcol = (System.Collections.IList)newcol;
+                }
+                else
+                {
+                    tarcol.Clear();
+                }
+                if (tarcol != null)
+                {
+
+
+                    foreach (var item in scol)
+                    {
+                        object newv = null;
+                        Copyref(item, ref newv);
+                        tarcol.Add(newv);
+                    }
+                    target = (T)tarcol;
+                }
+                else
+                {
+                    target = default(T);
+                }
+            }
+        }
+
+        /// <summary>
+        /// Copies to.
+        /// </summary>
+        /// <param name="target">The target.</param>
+        public void CopyTo(TSubClassType target)
+        {
+            foreach (var item in GetFieldNames())
+            {
+                var ctThis = GetValueContainer(item);
+                var ctTarget = target.GetValueContainer(item);
+                if (ctThis.IsCopyToAllowed)
+                {
+                    object temp = null;
+                    Copyref(this[item], ref temp);
+                    target[item] = temp;
+                }
+
+
+            }
+        }
+
+
+        /// <summary>
+        /// Occurs when [errors changed].
+        /// </summary>
+        public event EventHandler<DataErrorsChangedEventArgs> ErrorsChanged
+        {
+            add { _ErrorsChanged += value; }
+            remove { _ErrorsChanged -= value; }
+        }
+
+
+
+        /// <summary>
+        /// Gets the errors.
+        /// </summary>
+        /// <param name="propertyName">Name of the property.</param>
+        /// <returns>System.Collections.IEnumerable.</returns>
+        public System.Collections.IEnumerable GetErrors(string propertyName)
+        {
+            if (this.GetFieldNames().Contains(propertyName))
+            {
+                return this.GetValueContainer(propertyName).Errors;
+            }
+            else
+            {
+                return null;
+            }
+
+        }
+
+
+        /// <summary>
+        /// Gets a value indicating whether this instance has errors.
+        /// </summary>
+        /// <value><c>true</c> if this instance has errors; otherwise, <c>false</c>.</value>
+        //public bool HasErrors
+        //{
+        //    get
+        //    {
+        //        //  return false;
+        //        RefreshErrors();
+        //        return !string.IsNullOrEmpty(this.ErrorMessage);
+
+        //    }
+        //}
+
+
+        public bool HasErrors
+        {
+            get { return _HasErrorsLocator(this).Value; }
+            set { _HasErrorsLocator(this).SetValueAndTryNotify(value); }
+        }
+        #region Property bool HasErrors Setup        
+        protected Property<bool> _HasErrors = new Property<bool> { LocatorFunc = _HasErrorsLocator };
+        static Func<BindableBase, ValueContainer<bool>> _HasErrorsLocator = RegisterContainerLocator<bool>(nameof(HasErrors), model => model.Initialize(nameof(HasErrors), ref model._HasErrors, ref _HasErrorsLocator, _HasErrorsDefaultValueFactory));
+        static Func<bool> _HasErrorsDefaultValueFactory = () => false;
+        #endregion
+
+
+        /// <summary>
+        /// Refreshes the errors.
+        /// </summary>
+        protected virtual void OnGenrateErrorsMessage(IEnumerable<ErrorEntity> errors, StringBuilder errorMessageBuilder)
+        {
+            var sb = errorMessageBuilder;
+            var rt = GetAllErrors().Select(x =>
+            {
+                return sb.Append(x.PropertyName).Append("\t").Append(x.Message).Append("\t").AppendLine(x.Exception == null ? " " : x.Exception.ToString());
+            })
+                .ToArray();
+            this.SetErrorMessageAndTryNotify(sb.ToString());
+
+
+        }
+
+        /// <summary>
+        /// Gets all errors.
+        /// </summary>
+        /// <returns>ErrorEntity[].</returns>
+        public IEnumerable<ErrorEntity> GetAllErrors()
+        {
+            var errors = GetFieldNames()
+                 .SelectMany(name => this.GetValueContainer(name).Errors)
+                 .Where(x => x != null)
+                 .ToArray();
+            return errors;
+        }
+
+        //public override IDictionary<string,object >  Values
+        //{
+        //    get { return new BindableAccesser<TSubClassType>(this); }
+        //}
+        /// <summary>
+        /// 给这个模型分配的消息路由引用（延迟加载）
+        /// </summary>
+        /// <value>The event router.</value>
+
+
+        public override EventRouter LocalEventRouter
+        {
+            get { return _LocalEventRouterLocator(this).Value; }
+            set { _LocalEventRouterLocator(this).SetValueAndTryNotify(value); }
+        }
+        #region Property EventRouter LocalEventRouter Setup
+        protected Property<EventRouter> _LocalEventRouter = new Property<EventRouter> { LocatorFunc = _LocalEventRouterLocator };
+        static Func<BindableBase, ValueContainer<EventRouter>> _LocalEventRouterLocator = RegisterContainerLocator<EventRouter>("LocalEventRouter", model => model.Initialize("LocalEventRouter", ref model._LocalEventRouter, ref _LocalEventRouterLocator, _LocalEventRouterDefaultValueFactory));
+        static Func<EventRouter> _LocalEventRouterDefaultValueFactory = () => { return new EventRouter(); };
+        #endregion
+
+
+    }
+
 
     /// <summary>
     /// <para>Base type of bindable model.</para>
