+++ MVVMSidekick/MVVMSidekick.Shared/ViewModels/IViewModel1.cs
@@ -1,0 +1,81 @@
+﻿// ***********************************************************************
+// Assembly         : MVVMSidekick_Wp8
+// Author           : waywa
+// Created          : 05-17-2014
+//
+// Last Modified By : waywa
+// Last Modified On : 01-04-2015
+// ***********************************************************************
+// <copyright file="ViewModels.cs" company="">
+//     Copyright ©  2012
+// </copyright>
+// <summary></summary>
+// ***********************************************************************
+using System;
+using System.Threading.Tasks;
+#if NETFX_CORE
+
+
+#elif WPF
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Collections.Concurrent;
+using System.Windows.Navigation;
+using MVVMSidekick.Views;
+using System.Windows.Controls.Primitives;
+using MVVMSidekick.Utilities;
+using System.Windows.Threading;
+#elif SILVERLIGHT_5 || SILVERLIGHT_4
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Navigation;
+using System.Windows.Controls.Primitives;
+using System.Windows.Threading;
+#elif WINDOWS_PHONE_8 || WINDOWS_PHONE_7
+using System.Windows;
+using System.Windows.Controls;
+using Microsoft.Phone.Controls;
+using System.Windows.Data;
+using System.Windows.Navigation;
+using System.Windows.Controls.Primitives;
+using System.Windows.Threading;
+#endif
+
+
+
+
+
+
+namespace MVVMSidekick
+{
+
+    namespace ViewModels
+    {
+        /// <summary>
+        /// Interface IViewModel
+        /// </summary>
+        /// <typeparam name="TResult">The type of the t result.</typeparam>
+        public partial interface IViewModel<TResult> : IViewModel
+        {
+            /// <summary>
+            /// Waits for close with result.
+            /// </summary>
+            /// <param name="closingCallback">The closing callback.</param>
+            /// <returns>Task&lt;TResult&gt;.</returns>
+            Task<TResult> WaitForCloseWithResult(Action closingCallback = null);
+            /// <summary>
+            /// Gets or sets the result.
+            /// </summary>
+            /// <value>The result.</value>
+            TResult Result { get; set; }
+        }
+
+
+
+
+
+    }
+
+}
