+++ MVVMSidekick/MVVMSidekick_UAP10/MVVMSidekick_UAP10.csproj
@@ -116,6 +116,8 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Commands\" />
+    <Folder Include="Utilities\" />
+    <Folder Include="ViewModels\" />
   </ItemGroup>
   <Import Project="..\MVVMSidekick.Shared\MVVMSidekick.Shared.projitems" Label="Shared" />
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '14.0' ">
