+++ MVVMSidekick/Samples/Validation/project.json
@@ -5,7 +5,6 @@
     "Microsoft.Bcl.Async": "1.0.168",
     "Microsoft.Bcl.Build": "1.0.21",
     "Microsoft.NETCore.UniversalWindowsPlatform": "5.3.3",
-    "MVVM-Sidekick": "1.6.25",
     "Rx-Core": "2.2.5",
     "Rx-Interfaces": "2.2.5",
     "Rx-Linq": "2.2.5",
